+++ yast/database/__init__.py
@@ -1,0 +1,15 @@
+__all__ = [
+    "compile",
+    "DatabaseBackend",
+    "DatabaseSession",
+    "DatabaseTransaction",
+    "transaction",
+]
+
+from .core import (
+    DatabaseBackend,
+    DatabaseSession,
+    DatabaseTransaction,
+    compile,
+    transaction,
+)