+++ yast/middlewares/__init__.py
@@ -1,4 +1,5 @@
 __all__ = [
+    "Middleware",
     "ExceptionMiddleware",
     "CORSMiddleware",
     "GZipMiddleware",
@@ -9,10 +10,13 @@ __all__ = [
     "BaseHttpMiddleware",
     "ServerErrorMiddleware",
     "SessionMiddleware",
+    "DatabaseMiddleware",
 ]
 
 from .base import BaseHttpMiddleware
+from .core import Middleware
 from .cors import CORSMiddleware
+from .database import DatabaseMiddleware
 from .errors import ServerErrorMiddleware
 from .exception import ExceptionMiddleware
 from .gzip import GZipMiddleware