+++ changelog.md
@@ -7,6 +7,7 @@
 6. makefile: python coverage
 7. lifespan: as middleware
 8. graphql: context
+9. add DatabaseMiddleware: default postgresql
 
 # v0.2.3
 1. graphql: add graphiql